Implement Team service

+++ api-gateway/src/main/java/com/study/apigateway/dto/Team/response/TeamResponseDto.java
@@ -25,7 +25,7 @@ public class TeamResponseDto {
 
     private UUID creatorId;
 
-    private Integer totalMembers;
+    private Long totalMembers;
 
     private String avatarUrl;
 }
